+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 | :--------- | :------- | :---- |
 | Easy       | 21       | 492   |
 | Medium     | 56       | 837   |
-| Hard       | 21       | 328   |
+| Hard       | 22       | 328   |
 
 ## 0004 - Median Of Two Sorted Arrays
 
@@ -418,7 +418,7 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 - Two lists contains low and high
 - Combine two lists
 
-## 0086 - Scramble String
+## 0087 - Scramble String
 
 - Recursion
 - Pruning
@@ -465,6 +465,10 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 
 - Dynamic programming
 - $F_{n+1}=\sum_{i=0}^{n} F_i \times F_{n-i}$
+
+## 0097 - Interleaving String
+
+- Recursion
 
 ## 0098 - Validate Binary Search Tree